Split the code into bin/ and lib parts
Fixes #7.
